+++ tests/tensor/test_autograd.py
@@ -74,13 +74,13 @@ def test_trace_0():
     inds= [(0, 0, 0, 0), (0, 1, 0, 1), (0, 1, 1, 0), (1, 0, 0, 1), (1, 0, 1, 0), (1, 1, 1, 1)]
     vals= [0.25, -0.25, 0.5, 0.5, -0.25, 0.25]
     for t, val in zip(inds, vals):
-        H.A[()][t] = val
+        H[()][t] = val
     v = yast.Tensor(config=config_dense, s=(-1, -1))
     v.set_block(Ds=(2, 2), val='zeros')
     inds_v = [(0, 0), (0, 1), (1, 0), (1, 1)]
     vals_v = [0.1, 0.9, -0.9, 0.1]
     for t, val in zip(inds_v, vals_v):
-        v.A[()][t] = val
+        v[()][t] = val
 
     assert v.requires_grad is False
     v.requires_grad_()
@@ -99,14 +99,14 @@ def test_trace_0():
     g.set_block(Ds=(2, 2), val=expected_grad)
 
     for t in inds_v:
-        assert pytest.approx(v.A[()].grad[t].item(), rel=tol) == g.A[()][t].item()
+        assert pytest.approx(v.grad()[()][t].item(), rel=tol) == g[()][t].item()
 
 
 @pytest.mark.skipif(config_U1.backend.BACKEND_ID=="numpy", reason="numpy backend does not support autograd")
 def test_trace_1():
     H = yast.Tensor(config=config_U1, s=(-1, -1, 1, 1))
-    ta= [(-1, -1, -1, -1), (-1, 1, -1, 1), (-1, 1, 1, -1), (1, -1, -1, 1), (1, -1, 1, -1), (1, 1, 1, 1)]
-    ba= [0.25, -0.25, 0.5, 0.5, -0.25, 0.25]
+    ta = [(-1, -1, -1, -1), (-1, 1, -1, 1), (-1, 1, 1, -1), (1, -1, -1, 1), (1, -1, 1, -1), (1, 1, 1, 1)]
+    ba = [0.25, -0.25, 0.5, 0.5, -0.25, 0.25]
     for t, b in zip(ta, ba):
         H.set_block(ts=t, Ds=(1, 1, 1, 1), val=b)
     v = yast.Tensor(config=config_U1, s=(-1, -1, 1))
@@ -123,13 +123,13 @@ def test_trace_1():
     vHv = yast.tensordot(v, Hv, ((0, 1, 2), (0, 1, 2)), conj=(1, 0))
     vv = yast.vdot(v,v)
     vHv_vv = vHv / vv
-    loss= vHv_vv.to_number()
+    loss = vHv_vv.to_number()
 
     loss.backward()
-    expected_grad= [0.12046400951814398, -0.013384889946460365, 0.013384889946460365,\
+    expected_grad = [0.12046400951814398, -0.013384889946460365, 0.013384889946460365,\
         0.12046400951814398]
     for t, g in zip(tv, expected_grad):
-        assert pytest.approx(v.A[t].grad.item(), rel=tol) == g
+        assert pytest.approx(v.grad()[t].item(), rel=tol) == g
 
 if __name__ == '__main__':
     unittest.main()