+++ CHANGELOG.md
@@ -145,3 +145,4 @@
 - Transition to 1d data structure
 - `unique_dtype` replaced by `get_dtype`
 - new function '__setitem__()' that gives direct access to change existing blocks
+- in rsqrt(x, cutoff) cutoff is done with respect to x, not sqrt(x)